Read a Bruker OPUS spectrum binary file

Read single binary file acquired with an Bruker Vertex FTIR Instrument

read_opus_bin_univ(file_path, extract = c("spc"),
print_progress = TRUE, atm_comp_minus4offset = FALSE)

Arguments

file_path
Character vector with path to file

extract
Character vector of spectra types to extract from OPUS binary file. Default is c("spc"), which will extract the final spectra, e.g. expressed in absorbance (named AB in Bruker OPUS programs). Possible additional values for the character vector supplied to extract are "ScSm" (single channel spectrum of the sample measurement), \ code"ScRf" (single channel spectrum of the reference measurment), "IgSm" (interferogram of the sample measurment) and "IgRf" (interferogram of the reference measurement).

print_progress
Logical (default TRUE) whether a message is printed when an OPUS binary file is parsed into an R list entry.

atm_comp_minus4offset
Logical whether spectra after atmospheric compensation are read with an offset of -4 bites from Bruker OPUS files. Default is FALSE.